feat: move streaming output helps to command base class
